+++ components/restreamer/client/cypress/integration/check_streaming_state.js
@@ -6,72 +6,53 @@ describe('CHECK STREAMING STATE', () => {
     cy.runTestStream('rtmp://' + Cypress.env('host') + '/en/origin');
   });
 
-  it('Click Start All', () => {
+  it('1 Assert Start All', () => {
     cy.allOutputStart();
-    cy.wait(5000);
+    cy.wait(6000);
+    cy.checkStartedAllStated();
   });
 
-  const greenColor = 'rgb(50, 210, 150)';
-  const brownColor = 'rgb(122, 81, 40)';
+  it('2 Assert Stop All', () => {
+    cy.allOutputStop();
+    cy.wait(6000);
+    cy.checkStoppedAllStated();
+  });
 
-  it('Assert', () => {
-    cy.get(
-      '[data-icon="circle"][title="Serves failover live RTMP stream"]'
-    ).should('have.css', 'color', greenColor);
+  it('3 Click Start All', () => {
+    cy.allOutputStart();
+    cy.wait(6000);
+    cy.checkStartedAllStated();
+  });
 
-    cy.get(
-      '[data-icon="arrow-right"][title="Accepts main live RTMP stream"]'
-    ).should('have.css', 'color', greenColor);
+  it('4 Unselect sidechain', () => {
+    cy.get('[data-testid=Teamspeak]')
+      .parent()
+      .find("input[title='Sidechain']")
+      .first()
+      .click();
+    cy.wait(6000);
+  });
 
-    cy.get(
-      '[data-icon="arrow-right"][title="Accepts backup live RTMP stream"]'
-    ).should('have.css', 'color', greenColor);
+  it('4 Assert Started', () => {
+    cy.checkStartedAllStated();
+  });
 
-    cy.get(
-      '[data-icon="arrow-down"][title="Pulls origin live RTMP stream"]'
-    ).should('have.css', 'color', greenColor);
+  it('5 Select sidechain', () => {
+    cy.get('[data-testid=Teamspeak]')
+      .parent()
+      .find("input[title='Sidechain']")
+      .first()
+      .click();
+    cy.wait(6000);
+  });
 
-    cy.get('[data-testid=SINGLE] [data-icon="arrow-right"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
+  it('5 Assert Started', () => {
+    cy.checkStartedAllStated();
+  });
 
-    cy.get('[data-testid=RU] [data-icon="arrow-right"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
-
-    cy.wait(5000);
-    cy.get('[data-testid=Teamspeak] [data-icon="circle"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
-
-    cy.get('[data-testid=Twitter] [data-icon="circle"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
-
-    cy.get('[data-testid="[Manual Start] FB"] [data-icon="circle"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
-
-    cy.get('[data-testid="[Manual Start] YT"] [data-icon="dot-circle"]').should(
-      'have.css',
-      'color',
-      brownColor
-    );
-
-    cy.get('[data-testid="File Record"] [data-icon="circle"]').should(
-      'have.css',
-      'color',
-      greenColor
-    );
+  it('6 Assert Stopped', () => {
+    cy.allOutputStop();
+    cy.wait(6000);
+    cy.checkStoppedAllStated();
   });
 });