+++ components/restreamer/client/src/components/InputEndpoint.svelte
@@ -76,7 +76,7 @@
               class="fas fa-dot-circle"
               title="Serves {isFailover
                 ? 'failover '
-                : ''} live {endpoint.kind} stream"
+                : ''}live {endpoint.kind} stream"
             /></span
           >
         {:else}
@@ -85,7 +85,7 @@
               class="far fa-dot-circle"
               title="Serves {isFailover
                 ? 'failover '
-                : ''} live {endpoint.kind} stream"
+                : ''}live {endpoint.kind} stream"
             /></span
           >
         {/if}