+++ components/restreamer/src/ffmpeg/restreamer.rs
@@ -9,36 +9,58 @@ use std::{
 
 use ephyr_log::log;
 use futures::{future, pin_mut, FutureExt as _, TryFutureExt as _};
-use tokio::{process::Command, time};
+use tokio::{process::Command, sync::watch, time};
 
 use crate::{
     display_panic,
     ffmpeg::restreamer_kind::RestreamerKind,
     state::{State, Status},
-    types::DroppableAbortHandle,
 };
+
+/// Status of [Restreamer] process
+///
+/// Using for communication through [`tokio::sync::watch`]
+/// between [`Restreamer`] and [`MixingRestreamer`] with [`RestreamerKind`].
+///
+/// [`MixingRestreamer`]: crate::ffmpeg::MixingRestreamer
+#[derive(PartialEq, Eq, Debug, Clone, Copy)]
+pub(crate) enum RestreamerStatus {
+    /// [`Restreamer`] process is started and running
+    Started = 0,
+    /// [`Restreamer`] process is finishing
+    Finished = 1,
+}
 
 /// Handle to a running [FFmpeg] process performing a re-streaming.
 ///
 /// [FFmpeg]: https://ffmpeg.org
 #[derive(Debug)]
 pub struct Restreamer {
-    /// Abort handle of a spawned [FFmpeg] process of this [`Restreamer`].
-    ///
-    /// [FFmpeg]: https://ffmpeg.org
-    _abort: DroppableAbortHandle,
-
     /// Kind of a spawned [FFmpeg] process describing the actual job it
     /// performs.
     ///
     /// [FFmpeg]: https://ffmpeg.org
     pub kind: RestreamerKind,
+
+    /// Handle for stopping [FFmpeg] process of this [`Restreamer`].
+    ///
+    /// Kill with SIGTERM in normal scenario
+    ///
+    /// [FFmpeg]: https://ffmpeg.org
+    kill_tx: watch::Sender<RestreamerStatus>,
+
+    /// Handle for hanged [FFmpeg] process of this [`Restreamer`].
+    ///
+    /// Kill with SIGKILL if handed
+    ///
+    /// [FFmpeg]: https://ffmpeg.org
+    abort_if_hanged: future::AbortHandle,
 }
 
 impl Restreamer {
     /// Creates a new [`Restreamer`] spawning the actual [FFmpeg] process in
     /// background. Once this [`Restreamer`] is dropped, its [FFmpeg] process is
-    /// aborted.
+    /// killed with SIGTERM or aborted.
     ///
     /// [FFmpeg]: https://ffmpeg.org
     #[must_use]
@@ -50,11 +72,14 @@ impl Restreamer {
         let (kind_for_abort, state_for_abort) = (kind.clone(), state.clone());
         let kind_for_spawn = kind.clone();
         let mut time_of_fail: Option<DateTime<Utc>> = None;
+        let (kill_tx, kill_rx) = watch::channel(RestreamerStatus::Started);
 
-        let (spawner, abort_handle) = future::abortable(async move {
+        let (spawner, abort_if_hanged) = future::abortable(async move {
+            let kill_rx_for_loop = kill_rx.clone();
             loop {
                 let (kind, state) = (&kind_for_spawn, &state);
                 let mut cmd = Command::new(ffmpeg_path.as_ref());
+                let kill_rx_for_ffmpeg = kill_rx.clone();
 
                 let _ = AssertUnwindSafe(
                     async move {
@@ -80,7 +105,7 @@ impl Restreamer {
                         })
                         .await?;
 
-                        let running = kind.run_ffmpeg(cmd);
+                        let running = kind.run_ffmpeg(cmd, kill_rx_for_ffmpeg);
                         pin_mut!(running);
 
                         let set_online = async move {
@@ -123,18 +148,23 @@ impl Restreamer {
                     );
                 });
 
+                if *kill_rx_for_loop.borrow() == RestreamerStatus::Finished {
+                    break;
+                }
+
                 time::sleep(Duration::from_secs(2)).await;
             }
         });
 
-        // Spawn FFmpeg re-streamer as a child process.
+        // Spawn FFmpeg re-streamer manager as a child process.
         drop(tokio::spawn(spawner.map(move |_| {
             kind_for_abort.renew_status(Status::Offline, &state_for_abort);
         })));
 
         Self {
-            _abort: DroppableAbortHandle::new(abort_handle),
             kind,
+            kill_tx,
+            abort_if_hanged,
         }
     }
 
@@ -164,5 +194,22 @@ impl Restreamer {
                 kind.renew_status(new_status, state);
             }
         }
+    }
+}
+
+impl Drop for Restreamer {
+    /// Send signal that [`Restreamer`] process is finished
+    fn drop(&mut self) {
+        // Send notification to kill FFMPEG with SIGTERM
+        log::debug!("Send signal to FFmpeg's");
+        let _ = self.kill_tx.send(RestreamerStatus::Finished);
+
+        // If FFmpeg wasn't killed kill it with SIGKILL
+        let abort_for_future = self.abort_if_hanged.clone();
+        drop(tokio::spawn(async move {
+            log::debug!("Abort Restreamer in 5 sec if not killed");
+            time::sleep(Duration::from_secs(5)).await;
+            abort_for_future.abort();
+        }));
     }
 }