+++ components/restreamer/CHANGELOG.md
@@ -18,19 +18,27 @@ All user visible changes to this project will be documented in this file. This p
   - Allow to specify `identity` of Teamspeak mixer ([#6], [#39]);
   - Add `sidechain` option to `Mixin` in `Output` ([#70], [#203]).
 
+
+### Fixed
+- 'Gray screen' appears during switching from 'main' to 'backup' endpoint ([#164], [#204]).
+
+
 ### Miscellaneous
 - Server updates:
-  - `ffmpeg` from 4.4 to 5.1 ([e1faef9]);
-  - `SRS` server updated to v4.0-r1 ([e1faef9], [#200]);
+  - [FFmpeg] from 4.4 to 5.1 ([e1faef9]);
+  - [SRS] server updated to v4.0-r1 ([e1faef9], [#200]);
 
 - Update Tokio to v1+, Actix to v4+ and related libs ([#193]);
 - Split `ffmpeg.rs` to separate modules ([#202]);
 - Dockerfile image moved from CentOS 7 to Ubuntu 20.04 ([#200]);
 - Added test for check file recording ([#197]);
 - Use FIFO for feeding data into FFmpeg in mixer output ([#199]);
+- Gracefully shutdown [FFmpeg] processes ([#204]).
+
 
 [#6]: /../../issues/6
 [#70]: /../../issues/70
+[#164]: /../../issues/164
 
 [e1faef9]: /../../commit/e1faef91cc8551505afdf7fc4622c530f9e2c6f6
 [#39]: /../../pull/39
@@ -40,6 +48,7 @@ All user visible changes to this project will be documented in this file. This p
 [#200]: /../../pull/200
 [#202]: /../../pull/202
 [#203]: /../../pull/203
+[#204]: /../../pull/204
 
 
 
@@ -455,3 +464,6 @@ All user visible changes to this project will be documented in this file. This p
 [SRT]: https://en.wikipedia.org/wiki/Secure_Reliable_Transport
 [TeamSpeak]: https://teamspeak.com 
 [Ubuntu]: https://ubuntu.com
+[FFmpeg]: https://ffmpeg.org
+[SRS]: https://github.com/ossrs/srs
+